Add support to add a logfilter
To implement log sanitization according to KEP-1753 we need to be able
to install a LogFilter method that can replace the arguments to logging
statements to prevent accidental leaking of secrets, certificates or
passwords.

The interface of the filter defines three functions for the three kinds
of logging functions exposed by klog.
```
type LogFilter interface {
  Filter(args []interface{}) ([]interface{})
  FilterF(format string, args []interface{}) (string,[]interface{})
  FilterS(msg string, keysAndValues []interface{}) (string, []interface{})
}
```