klogr: add support for structured logging via klog
Now that klog has support for structured logging, it makes sense to
use that also through klogr. The log output without a structured
logger becomes more readable and key/value semantic is not lost
anymore with a structured logger.

However, revising the output handling is a breaking change for those
who depend on the traditional output. Therefore this is implemented in
a backward compatible way:
- code calling New() without parameters continues to work and
  gets the traditional behavior
- deferring the complete output handling to klog is enabled
  via a functional option